LayoutTests/fast/css-grid-layout/grid-item-order-in-content-sized-columns-resolution-expected.txt

PASS grid-template-columns is 57.78125px 20px after grid row swap.
PASS grid-template-columns is 73.328125px 50px after grid row swap.
PASS grid-template-columns is 57.78125px 15.546875px after grid row swap.
PASS grid-template-columns is 57.78125px 15.546875px after grid row swap.
PASS grid-template-columns is 73.328125px 0px after grid row swap.
PASS grid-template-columns is 57.78125px 15.546875px after grid row swap.
PASS grid-template-columns is 57.78125px 15.546875px after grid row swap.
PASS grid-template-columns is 57.78125px 15.546875px after grid row swap.
PASS successfullyParsed is true

TEST COMPLETE
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
